requirements.txt
flask
requests
python-dotenv
flask-login
flask-bcrypt
